refactor(estimates): Phase 4 — split NewEstimateForm into step components
NewEstimateForm.tsx: 1572 → 290 lines. Extracted Step1WhoAndWhat,
Step2Pricing, Step3Adjustments, Step4ReviewAndSend into
apps/web/app/app/estimates/new/components/.
